// src/types/player.ts
export interface Player {
  id: string;
  name: string;
  position: string;
  number: number;
  birthDate: string;
  height: number;
  weight: number;
  nationality: string;
  image: string;
  matches: number;
  goals: number;
  assists: number;
  yellowCards: number;
  redCards: number;
  teamId: string;
}